async_main
Despite std::execution's support for authoring asynchronous code in C++
the entry point to C++ programs remains fundamentally synchronous. main
is a synchronous function which:

- Completes successfully yielding a single integer (which, by
  convention, is zero on success, non-zero otherwise), and
- Can throw (which causes program termination)

This commit adds a truly asynchronous entry point for C++ programs:
async_main. This is used as follows:

  stdexec::sender auto async_main(std::span<const char* const> argv) {
    // ...
  }

  #include <exec/async_main.hpp>

The sender yielded by async_main is permitted to have the following
completions:

- stdexec::set_value_t(): Equivalent to returning zero from main
- stdexec::set_error_t(int): Equivalent to returning the sent integer
  (which is expected to be non-zero) from main
- stdexec::set_error_t(std::exception_ptr): Equivalent to throwing the
  sent exception from main

Note that this is not a straightforward translation of the completion
signatures of synchronous main. This is because synchronous main abuses
its value (i.e. return) channel by encapsulating therein both success
(returning zero) and error (returning non-zero). This difference
improves the composability of asynchronous main. Consider a program
which performs three tasks:

  stdexec::sender auto task_a();
  stdexec::sender auto task_b();
  stdexec::sender auto task_c();

  stdexec::sender auto async_main(std::span<const char* const> argv) {
    return stdexec::when_all(task_a(), task_b(), task_c());
  }

Now imagine one of the tasks encounters an error, and the program should
end (with some error code). Were async_main a straightforward port of
synchronous main we would expect the above to be accomplished via a
completion signature of the form stdexec::set_value_t(int). But if
task_a, _b, and _c had such a completion signature that would mean that:

- The completion signature of the stdexec::when_all sender wrapping all
  of them would be stdexec::set_value_t(int, int, int), which wouldn't
  be what is expected (i.e. the program wouldn't compile), and
- Despite morally being an error sending set_value with a non-zero value
  would be regarded by when_all as success, which would not cause it to
  stop the other children and thereby end the process

Note that unlike stdexec::sync_wait this implementation of async_main
makes no attempt to provide an execution context or scheduler. This is
because there is no universal choice which is suitable for all programs
and therefore this choice is deferred to the person providing
async_main.